+++ src/resqml2_0_1/IjkGridParametricRepresentation.cpp
@@ -840,10 +840,10 @@ void IjkGridParametricRepresentation::getXyzPointsOfBlockOfPatch(const unsigned 
 			filespace);
 
 		unsigned long long slab_size = 1;
-		for (unsigned int h = 0; h < 2; ++h) {
+		for (size_t h = 0; h < 2; ++h) {
 			slab_size *= blockSizeInEachDimension[h];
 		}
-		for (unsigned int h = 0; h < 2; ++h) {
+		for (size_t h = 0; h < 2; ++h) {
 			slab_size *= blockCountPerDimension[h];
 		}
